Added the new PACea_fetch.R function. This is the main function that the user will engage with. Once the desired datasets, with names fetch_names, has been found using the PACea_info() function, and once the desired regions have been found using the (coming soon) function PACea_regions(), the data can be downloaded across any desired year range and month range.
Note that a major feature of this function is that the data can either be returned as a .csv file or as an R data.frame object.

Another nice feature of the function, is that the data.frame will only return the data up to the resolution of the request. For example, if annual data are requested, then the data.frame will not be replicated over the months (i.e. no month column will be returned).

Futhermore, the function will 'pad' the data.frame with NA values accordingly. Thus, if data are unavailable for a given year/month combination, the entries will be returned as NA INSTEAD of dropping the rows. This is the required format for many R packages and analyses.

Finally, the function is written to enable efficient data storage. Coastwide indices can be stored efficiently without the need to replicate the values across all the Poly_IDs, reducing the storage by ~100 fold. Similarly, annual indices do not need to be stored as monthly indices, saving ~12 fold storage.