Added the PACea_regions() function. This is the last of the 3 main functions used by the typical user.
This function allows the user to search for, and plot, all the regions available in the PACea package. The names of each of the regions are provided to the user, for use in the PACea_fetch() function.

Note that ggplot2 is required to plot the regions. A warning message is given to the user about instructions to install ggplot2 (optional).

The sf polygons (with a sensible CRS) can also be returned to the user, if they so desire.